v0.0.5-alpha.3 massive refactor of the codebase
so every pre-release version up until now has been made redundant and has been ripped out.

Gamemode is now the core component, 'Event' - if it gets used at all - will be used to manage a Gamemode and that is it. In lieu of this, TrainingService can just do that instead. 'Event' would only be useful for tracking data, maybe.

anyway, rodux implementation should be easier when i write the training service methods and what not. 
after that then it's just checking rodux works as intended and then fleshing out promise functionality within gamemode.

then commandservice needs to be done to dispatch the store to fire these triggers and what not.